Total rewrite of bbex.js
Read the source of backbone.js, which helped to clarify some
of the things that I didn't understand.

Still some gaps, but a bit closer.

Would prefer for some of this code to be a bit more event driven,
but for now, it is doing the right things, so now to get it to
do those things correctly.