Refactor to make it cleaner to add to the collection.
This involved a long digression down fixing tsapp:

https://github.com/cdent/tsapp/commit/793b799256acfe17bbeceb4d69a576ef982bb63d

The goal with this change is for other code to create tiddlers
and add them to the model on the fly as needed (for example upon
messages from a websocket). This has resulted in a bit cleaner handling
in the views, but some confusion of responsibility between the views
used by the model and collection.

Apparently this is a common thing with backbone.